allow default SEMVER label to be configured
