+++ docs/pages/docs/configuration/autorc.mdx
@@ -243,6 +243,7 @@ You can customize everything about a label
 - `overwrite` - Overwrite the default label(s) associated with the `releaseType`. (default: `false`)
 - `changelogTitle` - The title to use in the changelog
 - `description` - The description to use when creating the label
+- `default` - Marks this label as the default label for unlabelled PRs (`patch` is the default "default label")
 - `color` - The color of the label. Can be specified as a string in any of [these](https://github.com/bgrins/TinyColor#accepted-string-input) ways. If not specified the color is random
 
 ```json
