+++ packages/core/src/__tests__/release.test.ts
@@ -1,11 +1,7 @@
 import Git from "../git";
 import LogParse from "../log-parse";
-import Release, {
-  defaultLabels,
-  getVersionMap,
-  ILabelDefinition,
-} from "../release";
-import SEMVER from "../semver";
+import Release, { getVersionMap } from "../release";
+import SEMVER, { defaultLabels, ILabelDefinition } from "../semver";
 import { dummyLog } from "../utils/logger";
 import makeCommitFromMsg from "./make-commit-from-msg";
 import child from "child_process";
