+++ packages/core/src/semver.ts
@@ -1,5 +1,12 @@
-import { VersionLabel } from "./release";
 import { ReleaseCalculationOptions } from "./types";
+import * as t from "io-ts";
+
+export type VersionLabel =
+  | SEMVER.major
+  | SEMVER.minor
+  | SEMVER.patch
+  | "skip"
+  | "release";
 
 enum SEMVER {
   major = "major",
@@ -21,6 +28,119 @@ export type IVersionLabels = Map<VersionLabel | "none", string[]>;
 
 export default SEMVER;
 
+export const releaseLabels: VersionLabel[] = [
+  SEMVER.major,
+  SEMVER.minor,
+  SEMVER.patch,
+  "skip",
+  "release",
+];
+
+/** Determine if a label is a label used for versioning */
+export const isVersionLabel = (label: string): label is VersionLabel =>
+  releaseLabels.includes(label as VersionLabel);
+
+const labelDefinitionRequired = t.type({
+  /** The label text */
+  name: t.string,
+});
+
+const labelDefinitionOptional = t.partial({
+  /** A title to put in the changelog for the label */
+  changelogTitle: t.string,
+  /** The color of the label */
+  color: t.string,
+  /** The description of the label */
+  description: t.string,
+  /** What type of release this label signifies */
+  releaseType: t.union([
+    t.literal("none"),
+    t.literal("skip"),
+    ...releaseLabels.map((l) => t.literal(l)),
+  ]),
+  /** Whether to overwrite the base label */
+  overwrite: t.boolean,
+  /** Marks this label as the default label for unlabelled PRs */
+  default: t.boolean,
+});
+
+export const labelDefinition = t.intersection([
+  labelDefinitionOptional,
+  labelDefinitionRequired,
+]);
+export type ILabelDefinition = t.TypeOf<typeof labelDefinition>;
+
+export const defaultLabels: ILabelDefinition[] = [
+  {
+    name: "major",
+    changelogTitle: "💥 Breaking Change",
+    description: "Increment the major version when merged",
+    releaseType: SEMVER.major,
+    color: "#C5000B",
+  },
+  {
+    name: "minor",
+    changelogTitle: "🚀 Enhancement",
+    description: "Increment the minor version when merged",
+    releaseType: SEMVER.minor,
+    color: "#F1A60E",
+  },
+  {
+    name: "patch",
+    changelogTitle: "🐛 Bug Fix",
+    description: "Increment the patch version when merged",
+    releaseType: SEMVER.patch,
+    color: "#870048",
+  },
+  {
+    name: "skip-release",
+    description: "Preserve the current version when merged",
+    releaseType: "skip",
+    color: "#bf5416",
+  },
+  {
+    name: "release",
+    description: "Create a release when this pr is merged",
+    releaseType: "release",
+    color: "#007f70",
+  },
+  {
+    name: "internal",
+    changelogTitle: "🏠 Internal",
+    description: "Changes only affect the internal API",
+    releaseType: "none",
+    color: "#696969",
+  },
+  {
+    name: "documentation",
+    changelogTitle: "📝 Documentation",
+    description: "Changes only affect the documentation",
+    releaseType: "none",
+    color: "#cfd3d7",
+  },
+  {
+    name: "tests",
+    changelogTitle: "🧪 Tests",
+    description: "Add or improve existing tests",
+    releaseType: "none",
+    color: "#ffd3cc",
+  },
+  {
+    name: "dependencies",
+    changelogTitle: "🔩 Dependency Updates",
+    description: "Update one or more dependencies version",
+    releaseType: "none",
+    color: "#8732bc",
+  },
+  {
+    name: "performance",
+    changelogTitle: "🏎 Performance",
+    description: "Improve performance of an existing feature",
+    releaseType: SEMVER.patch,
+    color: "#f4b2d8",
+  },
+];
+
 /** Given two labels determine the next SEMVER bump. */
 export function getHigherSemverTag(left: SEMVER, right: string): SEMVER {
   if (left === SEMVER.major || right === SEMVER.major) {
@@ -40,17 +160,25 @@ export function getHigherSemverTag(left: SEMVER, right: string): SEMVER {
  * strategy.
  */
 export function calculateSemVerBump(
-  labels: string[][],
+  prLabels: string[][],
   labelMap: IVersionLabels,
-  { onlyPublishWithReleaseLabel }: ReleaseCalculationOptions = {}
+  {
+    onlyPublishWithReleaseLabel,
+    labels = defaultLabels,
+  }: ReleaseCalculationOptions & {
+    /** The project's labels */
+    labels?: ILabelDefinition[];
+  } = {}
 ) {
+  const defaultLabel =
+    labels.find((l) => l.default)?.releaseType || SEMVER.patch;
   const labelSet = new Set<string>();
   const skipReleaseLabels = labelMap.get("skip") || [];
 
-  labels.forEach((pr, index) => {
+  prLabels.forEach((pr, index) => {
     // If the head pr has no labels we default to a patch
     if (pr.length === 0 && index === 0) {
-      labelSet.add(SEMVER.patch);
+      labelSet.add(defaultLabel);
     }
 
     pr.forEach((label) => {
@@ -64,7 +192,7 @@ export function calculateSemVerBump(
     });
   });
 
-  const lastMergedCommitLabels = labels[0] || [];
+  const lastMergedCommitLabels = prLabels[0] || [];
   const releaseLabels = labelMap.get("release") || [];
   const skipRelease = onlyPublishWithReleaseLabel
     ? !lastMergedCommitLabels.some((label) => releaseLabels.includes(label))
