+++ plugins/jira/__tests__/jira.test.ts
@@ -4,7 +4,7 @@ import Changelog, {
   IGenerateReleaseNotesOptions,
 } from "@auto-it/core/dist/changelog";
 import LogParse from "@auto-it/core/dist/log-parse";
-import { defaultLabels } from "@auto-it/core/dist/release";
+import { defaultLabels } from "@auto-it/core/dist/semver";
 import { dummyLog } from "@auto-it/core/dist/utils/logger";
 import {
   makeChangelogHooks,
